Move selectors to separate file for readability
